Fixed arm transforms
All the servo calculations were correct – it was just the transforms in
drawing the model that weren’t working.
I changed the transforms so that they are all in a parent-child
hierarchy, so SetJointsTransforms was simpler.
I removed the scaling of the transforms too, since they propagate
through the hierarchy.
Didn’t bother with trying to do the scaling a different way to get the
elongated blocks.
Also made a minor change to InitializeScreen because I couldn’t get the
expected results when I changed the tilt of the screen.